// src/modules/configuration/model/json-property.interface.ts
export interface JsonProperty {
    IDJsonProperty: number;
    IDJsonService: number;
    State: number;
    PropertyPath: string;
    Since: string;
    PropertyClass: string;
    TypeName: string;
    PropertyValue: string;
    DataBeanClass: any;
    PropertyIndex: number;
    UserValue: string;
    BPMDataBeanProperty: string;
    BPMDataBeanClass: string;
    Type: number;
    DocumentType: number;
    IDForm: number;
    IDFormClass: number;
    IDJsonServiceClass: number;
    IDFormSrc: number;
    IDJsonServiceSrc: number;
}